+++ CHANGELOG.md
@@ -6,8 +6,11 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [x.x.x]
 ### Added
-- Extend the delete variants command to include analysis date, track and institute
+- Extend the delete variants command to print analysis date, track, institute, status and research status
+- Delete variants by type of analysis (wgs|wes|panel)
+- Links to cBioPortal, MutanTP53, IARC TP53, OncoKB, MyCancerGenome, CIViC
 ### Fixed
+- Deleted variants count
 ### Changed
 - Print output of variants delete command as a tab separated table